Use long int for squared quantities.

+++ L1Trigger/L1THGCal/interface/backend_emulator/HGCalCluster_SA.h
@@ -52,16 +52,16 @@ namespace l1thgcfirmware {
     void set_e_h_early( unsigned int e_h_early ) { e_h_early_ = e_h_early; }
     void set_w( unsigned int w ) { w_ = w; }
     void set_n_tc_w( unsigned int n_tc_w ) { n_tc_w_ = n_tc_w; }
-    void set_w2( unsigned int w2 ) { w2_ = w2; }
+    void set_w2( unsigned long int w2 ) { w2_ = w2; }
     void set_wz( unsigned int wz ) { wz_ = wz; }
     void set_weta( unsigned int weta ) { weta_ = weta; }
     void set_wphi( unsigned int wphi ) { wphi_ = wphi; }
     void set_wroz( unsigned int wroz ) { wroz_ = wroz; }
-    void set_wz2( unsigned int wz2 ) { wz2_ = wz2; }
-    void set_weta2( unsigned int weta2 ) { weta2_ = weta2; }
-    void set_wphi2( unsigned int wphi2 ) { wphi2_ = wphi2; }
-    void set_wroz2( unsigned int wroz2 ) { wroz2_ = wroz2; }
-    void set_layerbits( unsigned int layerbits ) { layerbits_ = layerbits; }
+    void set_wz2( unsigned long int wz2 ) { wz2_ = wz2; }
+    void set_weta2( unsigned long int weta2 ) { weta2_ = weta2; }
+    void set_wphi2( unsigned long int wphi2 ) { wphi2_ = wphi2; }
+    void set_wroz2( unsigned long int wroz2 ) { wroz2_ = wroz2; }
+    void set_layerbits( unsigned long int layerbits ) { layerbits_ = layerbits; }
     void set_sat_tc( bool sat_tc ) { sat_tc_ = sat_tc; }
     void set_shapeq( unsigned int shapeq ) { shapeq_ = shapeq; }
 
@@ -78,16 +78,16 @@ namespace l1thgcfirmware {
     unsigned int e_h_early() const { return e_h_early_; }
     unsigned int w() const { return w_; }
     unsigned int n_tc_w() const { return n_tc_w_; }
-    unsigned int w2() const { return w2_; }
+    unsigned long int w2() const { return w2_; }
     unsigned int wz() const { return wz_; }
     unsigned int weta() const { return weta_; }
     unsigned int wphi() const { return wphi_; }
     unsigned int wroz() const { return wroz_; }
-    unsigned int wz2() const { return wz2_; }
-    unsigned int weta2() const { return weta2_; }
-    unsigned int wphi2() const { return wphi2_; }
-    unsigned int wroz2() const { return wroz2_; }
-    unsigned int layerbits() const { return layerbits_; }
+    unsigned long int wz2() const { return wz2_; }
+    unsigned long int weta2() const { return weta2_; }
+    unsigned long int wphi2() const { return wphi2_; }
+    unsigned long int wroz2() const { return wroz2_; }
+    unsigned long int layerbits() const { return layerbits_; }
     bool sat_tc() const { return sat_tc_; }
     unsigned int shapeq() const { return shapeq_; }
 
@@ -108,16 +108,16 @@ namespace l1thgcfirmware {
     unsigned int e_h_early_;
     unsigned int w_;
     unsigned int n_tc_w_;
-    unsigned int w2_;
+    unsigned long int w2_;
     unsigned int wz_;
     unsigned int weta_;
     unsigned int wphi_;
     unsigned int wroz_;
-    unsigned int wz2_;
-    unsigned int weta2_;
-    unsigned int wphi2_;
-    unsigned int wroz2_;
-    unsigned int layerbits_;
+    unsigned long int wz2_;
+    unsigned long int weta2_;
+    unsigned long int wphi2_;
+    unsigned long int wroz2_;
+    unsigned long int layerbits_;
     bool sat_tc_;
     unsigned int shapeq_;
